Apply texture to object
Also remove old triangles and replace sail with stb_image. I though sail
was the culprit for the image not rendering correctly, but it was
actually setting the texture to GL_RGB when passing in a png with
transparency. Still keeping stb_image instead because sail was painful
to set up, so changes to the dev environment might break it.