Refactor setup into Game class
Capturing lambdas are a bit unsafe to store, so to discourage that, use
an interface class instead. This required rewriting the existing setup
code to use a class instead of being in main(). (This class is very mess
and will need to be refactored sooner rather rather than later, but it
works for now.)